Pull out ?? and "Remind me" links into their own component
